restructure and add custom file support

+++ src/emulators/dolphin.py
@@ -28,8 +28,15 @@ class Dolphin:
         time.sleep(2)
         os.remove(zip_path)
         
-    
-
+    def verify_dolphin_zip(self, path_to_archive):
+        try:
+            with ZipFile(path_to_archive, 'r') as archive:
+                if 'Dolphin.exe' in archive.namelist():
+                    return True 
+                else:
+                    return False
+        except Exception:
+            return False
     def download_release(self, release):
         download_folder = os.getcwd()
         
@@ -41,40 +48,39 @@ class Dolphin:
         download_result = download_through_stream(response, download_path, progress_frame, 1024*203)
         progress_frame.destroy() 
         return download_result
-
         
-
-        
-            
-                          
-        
-    def install_dolphin_handler(self, updating=False):
-            
-        release_result = get_resources_release("Dolphin", headers=get_headers(self.settings.app.token))
-        if not all(release_result):
-            messagebox.showerror("Install Dolphin", f"There was an error while attempting to fetch the latest release of Dolphin:\n\n{release_result[1]}")
-            return 
-        release = release_result[1]
-        release.version = release.name.replace("Dolphin.", "").replace(".zip", "")
-        if release.version == self.metadata.get_installed_version("dolphin"):
-            if updating:
+    def install_dolphin_handler(self, updating=False, path_to_archive=None):
+        release_archive = path_to_archive
+        if path_to_archive is None:
+            release_result = get_resources_release("Dolphin", headers=get_headers(self.settings.app.token))
+            if not all(release_result):
+                messagebox.showerror("Install Dolphin", f"There was an error while attempting to fetch the latest release of Dolphin:\n\n{release_result[1]}")
                 return 
-            if not messagebox.askyesno("Dolphin", "You already have the latest version of Dolphin installed, install anyways?"):
-                return
-        download_result = self.download_release(release)
-        if not all(download_result):
-            if download_result[1] != "Cancelled":
-                messagebox.showerror("Error", f"There was an error while attempting to download the latest release of Dolphin\n\n{download_result[1]}")
+            release = release_result[1]
+            release.version = release.name.replace("Dolphin.", "").replace(".zip", "")
+            if release.version == self.metadata.get_installed_version("dolphin"):
+                if updating:
+                    return 
+                if not messagebox.askyesno("Dolphin", "You already have the latest version of Dolphin installed, install anyways?"):
+                    return
+            download_result = self.download_release(release)
+            if not all(download_result):
+                if download_result[1] != "Cancelled":
+                    messagebox.showerror("Error", f"There was an error while attempting to download the latest release of Dolphin\n\n{download_result[1]}")
+                return 
+            release_archive = download_result[1]
+        elif not self.verify_dolphin_zip(path_to_archive):
+            messagebox.showerror("Error", "The dolphin archive you have provided is invalid. ")
             return 
-        release_archive = download_result[1]
         extract_result = self.extract_release(release_archive)
-        if self.settings.app.delete_files == "True":
+        if path_to_archive is None and self.settings.app.delete_files == "True":
             os.remove(release_archive)
         if not all(extract_result):
             if extract_result[1]!="Cancelled":
                 messagebox.showerror("Extract Error", f"An error occurred while extracting the release: \n\n{extract_result[1]}")
             return 
-        self.metadata.update_installed_version("dolphin", release.version)
+        if path_to_archive is None:
+            self.metadata.update_installed_version("dolphin", release.version)
         messagebox.showinfo("Install Dolphin", f"Dolphin was successfully installed to {self.settings.dolphin.install_directory}")
